+++ apps/mobile/src/components/posts/video/status.tsx
@@ -1,5 +1,3 @@
-import { useEventListener } from 'expo'
-import { type VideoPlayer } from 'expo-video'
 import { View } from 'react-native'
 import Animated, {
   cancelAnimation,
@@ -8,6 +6,7 @@ import Animated, {
   withTiming,
 } from 'react-native-reanimated'
 import { StyleSheet } from 'react-native-unistyles'
+import { useEvent, type VideoPlayer } from 'react-native-video'
 
 type Props = {
   player: VideoPlayer
@@ -17,7 +16,7 @@ export function VideoStatus({ player }: Props) {
   const current = useSharedValue(0)
   const buffered = useSharedValue(0)
 
-  useEventListener(player, 'timeUpdate', (event) => {
+  useEvent(player, 'onProgress', (event) => {
     const nextCurrent = (event.currentTime / player.duration) * 100
 
     if (current.get() !== nextCurrent) {
@@ -30,7 +29,7 @@ export function VideoStatus({ player }: Props) {
       )
     }
 
-    const nextBuffered = (event.bufferedPosition / player.duration) * 100
+    const nextBuffered = (event.bufferDuration / player.duration) * 100
 
     if (buffered.get() !== nextBuffered) {
       cancelAnimation(buffered)